3rd_party/leaderboard/fail2drive/team_code/requirements.txt
carla==0.9.15
dictor==0.1.12
diskcache==5.4.0
einops==0.4.1
ephem==4.1.5
filterpy==1.4.5
gym==0.17.2
h5py==3.7.0
imgaug==0.4.0
jsonpickle==3.0.3
laspy==2.5.4
laszip==0.2.3
lxml==5.1.0
matplotlib==3.5.3
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.6.0.66
pexpect==4.9.0
Pillow==10.2.0
psutil==5.9.8
py-trees==0.8.3
pygame==2.6.0
pylint==2.15.0
pytictoc==1.5.3
rdp==0.8
requests==2.31.0
scikit-image==0.19.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==59.8.0
shapely==2.0.4
simple_watchdog_timer==0.1.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.14.0
tensorboardX==2.6.2.2
timm==1.0.11
torch==2.5.0
torchaudio==2.5.0
torchmetrics==0.11.0
torchvision==0.20.0
tqdm==4.66.1
transformers==4.21.2
transforms3d==0.4.1
ujson==5.9.0
wandb==0.16.6
webcolors==1.13
xmlschema==1.0.18
pandas
pyQt6